move getRFC3339Date to lib/common/utils
