+++ __tests__/unit/mocked-csn.test.js
@@ -14,7 +14,7 @@ describe("Tests for ORD document generated out of mocked csn files", () => {
         // Mock the createAuthConfig to return a default open configuration
         mockCreateAuthConfig(authentication);
 
-        jest.spyOn(require("../../lib/date"), "getRFC3339Date").mockReturnValue("2024-11-04T14:33:25+01:00");
+        jest.spyOn(require("../../lib/common/utils"), "getRFC3339Date").mockReturnValue("2024-11-04T14:33:25+01:00");
         ord = require("../../lib/ord");
     });
