+++ __tests__/unit/split-up-packages-csn.test.js
@@ -17,7 +17,7 @@ describe("ORD Generation for Business Accelerator Hub", () => {
             accessStrategies: [{ type: ORD_ACCESS_STRATEGY.Open }],
         });
 
-        jest.spyOn(require("../../lib/date"), "getRFC3339Date").mockReturnValue("2024-11-04T14:33:25+01:00");
+        jest.spyOn(require("../../lib/common/utils"), "getRFC3339Date").mockReturnValue("2024-11-04T14:33:25+01:00");
         // Require ord module after mocks are set up
         ord = require("../../lib/ord");
     });
