+++ lib/configuration.js
@@ -4,7 +4,7 @@ const _ = require("lodash");
 const { readFileSync } = require("fs");
 
 const Logger = require("./logger");
-const { getRFC3339Date } = require("./date");
+const { getRFC3339Date } = require("./common/utils");
 const defaults = require("./defaults");
 const { createEntityTypeMappingsItemTemplate } = require("./templates");
 const { CONTENT_MERGE_KEY, CDS_ELEMENT_KIND } = require("./constants");
